backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
annotated-types
anyio
asyncpg
attrs
banks
certifi
charset-normalizer
click
colorama
coloredlogs
dataclasses-json
Deprecated
deprecation
dirtyjson
distro
python-dotenv
duckdb
fastapi
filelock
filetype
flatbuffers
frozenlist
fsspec
gotrue
greenlet
griffe
h11
h2
hpack
httpcore
httpx
huggingface-hub
humanfriendly
hyperframe
idna
iniconfig
Jinja2
jiter
joblib
llama-index-core
llama-index-embeddings-huggingface
llama-index-embeddings-huggingface-optimum
llama-index-embeddings-ollama
llama-index-embeddings-openai
llama-index-instrumentation
llama-index-llms-ollama
llama-index-storage-chat-store-postgres
llama-index-vector-stores-duckdb
llama-index-workflows
Markdown
MarkupSafe
marshmallow
mpmath
multidict
mypy_extensions
nest-asyncio
networkx
nltk
numpy
ollama
onnx
onnxruntime
openai
optimum
packaging
pillow
platformdirs
pluggy
postgrest
propcache
protobuf
psycopg
psycopg-binary
psycopg-pool
pyarrow
pydantic
pydantic_core
Pygments
PyJWT
pyreadline3
pytest
pytest-mock
python-dateutil
python-multipart
PyYAML
realtime
regex
requests
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
six
sniffio
SQLAlchemy
starlette
storage3
StrEnum
supabase
supafunc
sympy
tenacity
threadpoolctl
tiktoken
timm
tokenizers
torch
torchvision
tqdm
transformers
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
websockets
wrapt
yarl
PyPDF2